requirements.txt
# StreamGuard - Production Dependencies
# Python 3.12
# Last Updated: March 21, 2026
# Tested with these exact versions - DO NOT update without testing

# Core dependencies
pydantic==2.12.5
python-dotenv==1.0.1

# Layer 1: PII Detection (Presidio)
presidio-analyzer==2.2.362
presidio-anonymizer==2.2.362
spacy==3.8.11

# Layer 2a & 2b: ML/HuggingFace (PyTorch CPU)
torch==2.10.0+cpu
transformers==5.3.0
safetensors==0.4.5
huggingface-hub==0.26.5

# Layer 3: Content Moderation
httpx[http2]==0.28.1

# Layer 4: Stateful Analysis (OPTIONAL)
# Uncomment the line below to enable multi-turn conversation tracking
# upstash-redis==1.7.0

# Additional dependencies
numpy==1.26.4
filelock==3.18.0
fsspec==2024.10.0
packaging==24.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
tokenizers==0.21.0
tqdm==4.67.1
typing-extensions==4.12.2

# PyTorch CPU installation
--extra-index-url https://download.pytorch.org/whl/cpu
